requirements.txt
# Core dependencies
fastapi==0.109.1
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-dotenv==1.0.0
pydantic==1.10.13
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9  # Required for PostgreSQL

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Utilities
email-validator==2.1.0.post1
python-dateutil==2.8.2
python-multipart==0.0.9

# API Documentation
description==0.1.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0

# CORS
starlette==0.36.3

# Logging
structlog==24.1.0

# Environment
python-dotenv==1.0.0

# Required for production
anyio==4.2.0
click==8.1.7
h11==0.14.0
idna==3.6
sniffio==1.3.0
typing-extensions==4.9.0
